ingestion/water_metrics_loader/requirements.txt
confluent-kafka
jsonschema
pyyaml
httpx
cachetools
authlib
